# Conflicts:
#	app/controllers/ActualizationController.java
#	app/controllers/CompilationLibrariesController.java
#	app/models/compiler/Board.java
#	app/models/compiler/Version_Object.java
#	app/models/project/global/Project.java
#	app/utilities/Server.java
#	app/utilities/hardware_updater/Master_Updater.java

+++ app/controllers/NotificationController.java
@@ -1,6 +1,7 @@
 package controllers;
 
 import com.avaje.ebean.Query;
+import com.google.inject.Inject;
 import io.swagger.annotations.*;
 import models.compiler.Board;
 import models.compiler.Version_Object;
@@ -17,25 +18,32 @@ import play.mvc.Controller;
 import play.mvc.Result;
 import play.mvc.Security;
 import utilities.Server;
+import utilities.enums.Notification_action;
 import utilities.enums.Notification_importance;
 import utilities.enums.Notification_level;
 import utilities.loggy.Loggy;
 import utilities.loginEntities.Secured_API;
+import utilities.notifications.Notification_Handler;
 import utilities.response.GlobalResult;
-import utilities.response.response_objects.Result_Unauthorized;
-import utilities.response.response_objects.Result_ok;
+import utilities.response.response_objects.*;
 import utilities.swagger.documentationClass.Swagger_B_Program_Version_New;
+import utilities.swagger.documentationClass.Swagger_Notification_Confirm;
 import utilities.swagger.documentationClass.Swagger_Notification_Read;
 import utilities.swagger.documentationClass.Swagger_Notification_Test;
 import utilities.swagger.outboundClass.Filter_List.Swagger_Notification_List;
 import utilities.swagger.outboundClass.Swagger_B_Program_Version;
 import utilities.swagger.outboundClass.Swagger_C_Program_Version;
+import utilities.swagger.outboundClass.Swagger_Notification_Button;
 import utilities.webSocket.WS_Becki_Website;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Api(value = "Not Documented API - InProgress or Stuck")
 public class NotificationController extends Controller {
+
+  @Inject
+  ProgramingPackageController programingPackageController;
 
   //####################################################################################################################
   static play.Logger.ALogger logger = play.Logger.of("Loggy");
@@ -43,7 +51,7 @@ public class NotificationController extends Controller {
   private static void send_notification(Person person, Notification notification) {
 
     // Pokud je notification_importance vyšší než "low" notifikaci uložím
-    if(!(notification.notification_importance == Notification_importance.low))
+    if((notification.notification_importance != Notification_importance.low)&&(notification.id == null))
       notification.save_object();
 
     // Pokud je uživatel přihlášený pošlu notifikaci přes websocket
@@ -56,6 +64,7 @@ public class NotificationController extends Controller {
   // Tvroba objektů jednotlivých notifikací ############################################################################
 
 
+  // Hotovo
   public static void starting_of_compilation(Person person, Version_Object version_object){
 
         Notification notification = new Notification(Notification_importance.low, Notification_level.info, person)
@@ -66,10 +75,11 @@ public class NotificationController extends Controller {
   }
 
   public static void upload_firmware_progress(Person person, String version_object){
-    // TODO
+    // TODO a taky zařadit pod objekt
   }
 
 
+  // Hotovo
   public static void successful_compilation(Person person, Version_Object version_object ){
 
       Notification notification = new Notification(Notification_importance.low, Notification_level.success, person)
@@ -80,6 +90,7 @@ public class NotificationController extends Controller {
       send_notification(person, notification);
   }
 
+  // Hotovo
   public static void unsuccessful_compilation_warn(Person person, Version_Object version_object, String reason){
 
       Notification notification = new Notification(Notification_importance.normal,  Notification_level.warning , person)
@@ -91,6 +102,7 @@ public class NotificationController extends Controller {
       send_notification(person, notification);
   }
 
+  // Hotovo
   public static void unsuccessful_compilation_error(Person person, Version_Object version_object, String result){
 
     Notification notification = new Notification(Notification_importance.normal, Notification_level.error, person)
@@ -102,6 +114,7 @@ public class NotificationController extends Controller {
     send_notification(person, notification);
   }
 
+  // Hotovo
   public static void upload_Instance_start(Person person, Homer_Instance instance){
 
     Notification notification = new Notification(Notification_importance.low,  Notification_level.info, person)
@@ -113,6 +126,7 @@ public class NotificationController extends Controller {
     send_notification(person, notification);
   }
 
+  // Hotovo
   public static void upload_Instance_was_successful(Person person, Homer_Instance instance){
 
     Notification notification = new Notification(Notification_importance.low, Notification_level.success, person)
@@ -124,7 +138,7 @@ public class NotificationController extends Controller {
     send_notification(person, notification);
   }
 
-
+  // Hotovo
   public static void upload_Instance_was_unsuccessful(Person person, Homer_Instance instance, String reason){
 
 
@@ -139,6 +153,7 @@ public class NotificationController extends Controller {
 
   }
 
+  // TODO zařadit pod objekt
   public static void upload_of_Instance_was_unsuccessful_with_error(Person person, Version_Object version_object){
 
     Notification notification = new Notification(Notification_importance.normal, Notification_level.error, person)
@@ -151,6 +166,7 @@ public class NotificationController extends Controller {
     send_notification(person, notification);
   }
 
+  // Hotovo
   public static void new_actualization_request_with_file(Person person, Board board){
 
     Notification notification = new Notification(Notification_importance.low, Notification_level.info, person)
@@ -163,6 +179,7 @@ public class NotificationController extends Controller {
 
   }
 
+  // Hotovo
   public static void new_actualization_request_on_version(Person person, Version_Object version_object){
 
     Notification notification = new Notification(Notification_importance.low, Notification_level.info, person)
@@ -176,6 +193,7 @@ public class NotificationController extends Controller {
 
   }
 
+  // Hotovo
   public static void new_actualization_request_homer_instance(Project project, Homer_Instance homer_instance){
 
     for(Person person : project.ownersOfProject) {
@@ -189,7 +207,7 @@ public class NotificationController extends Controller {
 
   }
 
-
+  // Hotovo
   public static void board_connect(Person person, Board board){
 
       Notification notification = new Notification(Notification_importance.low, Notification_level.info, person)
@@ -201,6 +219,7 @@ public class NotificationController extends Controller {
 
   }
 
+  // Hotovo
   public static void board_disconnect(Person person, Board board){
 
     Notification notification = new Notification(Notification_importance.low, Notification_level.info, person)
@@ -212,19 +231,19 @@ public class NotificationController extends Controller {
 
   }
 
-
+  //Hotovo
   public static void project_invitation(Person owner, Person receiver, Project project, Invitation invitation){
 
     Notification notification = new Notification(Notification_importance.normal, Notification_level.info, receiver)
-            .setText("User")
-            .setObject(Person.class, owner.id, owner.full_name, "")
+            .setText("User ")
+            .setObject(Person.class, owner.id, owner.full_name, null)
             .setText("wants to invite you into the project ")
             .setObject(Project.class, project.id, project.name, project.id)
             .setText(".")
             .setText("Do you agree?")
-            .setLink_ToTyrion("Yes", Server.tyrion_serverAddress + "/project/project/addParticipant/" + invitation.id + "/true")
+            .setLink("Yes", Server.tyrion_serverAddress + "/project/project/addParticipant/" + invitation.id + "/true")
             .setText(" / ")
-            .setLink_ToTyrion("No", Server.tyrion_serverAddress + "/project/project/addParticipant/" + invitation.id + "/false")
+            .setLink("No", Server.tyrion_serverAddress + "/project/project/addParticipant/" + invitation.id + "/false")
             .setText(".");
 
     invitation.notification_id = notification.id;
@@ -234,7 +253,7 @@ public class NotificationController extends Controller {
 
   }
 
-
+  //Hotovo
   public static void project_accepted_by_invited_person(Person owner, Person person, Project project){
 
     Notification notification = new Notification(Notification_importance.normal, Notification_level.info, owner)
@@ -248,6 +267,7 @@ public class NotificationController extends Controller {
 
   }
 
+  //Hotovo
   public static void project_rejected_by_invited_person(Person owner, Person person, Project project){
 
     Notification notification = new Notification(Notification_importance.normal, Notification_level.info, owner)
@@ -261,7 +281,7 @@ public class NotificationController extends Controller {
 
   }
 
-  public static void test_notification(Person person, String level, String importance, boolean confirmation_required){
+  public static void test_notification(Person person, String level, String importance, String type, String buttons){
 
     Notification_level lvl;
 
@@ -282,17 +302,66 @@ public class NotificationController extends Controller {
       default: lvl = Notification_level.info;break;
     }
 
-    Notification notification = new Notification(imp, lvl, person)
-            .setText("Test object: ")
-            .setObject(Person.class, person.id, person.full_name, "")
-            .setText("test bold text: ")
-            .setBoldText("bold text")
-            .setText("test link:")
-            .setLink_ToTyrion("TestLink","#");
+    Notification notification;
 
-    if(confirmation_required) notification.confirmation_required = true;
+    switch (type){
+      case "1":{
+        notification = new Notification(imp, lvl, person)
+              .setText("Test object: ")
+              .setObject(Person.class, person.id, person.full_name, null)
+              .setText(" test bold text: ")
+              .setBoldText("bold text ")
+              .setText("test link: ")
+              .setLink("TestLink","#");
+        break;}
+      case "2":{
+        notification = new Notification(imp, lvl, person)
+                .setText("Test object and long text: ")
+                .setObject(Person.class, person.id, person.full_name, null)
+                .setText(" test text: Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. ");
+        break;}
+      case "3":{
+        notification = new Notification(imp, lvl, person)
+                .setText("Test short text with link: ")
+                .setLink("TestLink","#");
+        break;}
+      case "4": {
+        notification = new Notification(imp, lvl, person)
+                .setText("Test object and link: ")
+                .setObject(Person.class, person.id, person.full_name, null)
+                .setText(" test link: ")
+                .setLink("Yes","#");
+        break;}
+      default:{
+        notification = new Notification(imp, lvl, person)
+                .setText("Test object: ")
+                .setObject(Person.class, person.id, person.full_name, null)
+                .setText(" test bold text: ")
+                .setBoldText("bold text ")
+                .setText("test link: ")
+                .setLink("TestLink","#");
+        break;}
+    }
+    switch (buttons){
+      case "0": break;
+      case "1":{
+        notification.setButton(Notification_action.confirm_notification, "test", "blue", "OK", false, false, false);
+        break;}
+      case "2":{
+        notification.setButton(Notification_action.accept_project_invitation, "test", "green", "Yes", false, false, true);
+        notification.setButton(Notification_action.reject_project_invitation, "test", "red", "No", false, false, true);
+        break;}
+      case "3":{
+        notification.setButton(Notification_action.accept_project_invitation, "test", "green", "Yes", true, false, false);
+        notification.setButton(Notification_action.reject_project_invitation, "test", "red", "No", true, false, false);
+        notification.setButton(Notification_action.confirm_notification, "test", "white", "Close", false, true, false);
+        break;}
+      default:{
+        notification.setButton(Notification_action.accept_project_invitation, "test", "green", "Yes", false, false, false);
+        break;}
+    }
 
-    send_notification(person,notification);
+    notification.send(person);
   }
 
 
@@ -348,15 +417,19 @@ public class NotificationController extends Controller {
           code = 200
   )
   @ApiResponses(value = {
-          @ApiResponse(code = 200, message = "Delete Successful",        response = Result_ok.class),
+          @ApiResponse(code = 200, message = "Delete Successful",       response = Result_ok.class),
+          @ApiResponse(code = 400, message = "Objects not found",       response = Result_NotFound.class),
           @ApiResponse(code = 401, message = "Unauthorized request",    response = Result_Unauthorized.class),
+          @ApiResponse(code = 403, message = "Need required permission",response = Result_PermissionRequired.class),
           @ApiResponse(code = 500, message = "Server side Error")
   })
   @Security.Authenticated(Secured_API.class)
   public Result delete_notification(@ApiParam(value = "notification_id String path", required = true) String notification_id){
     try {
 
-      Notification notification =  Notification.find.byId(notification_id);
+      Notification notification = Notification.find.byId(notification_id);
+      if (notification == null) return GlobalResult.notFoundObject("Notification does not exist");
+
       if( !notification.delete_permission()) return GlobalResult.forbidden_Permission();
 
       notification.delete();
@@ -387,6 +460,7 @@ public class NotificationController extends Controller {
   )
   @ApiResponses(value = {
           @ApiResponse(code = 200, message = "Successfully marked as read", response = Result_ok.class),
+          @ApiResponse(code = 400, message = "Some Json value Missing",     response = Result_JsonValueMissing.class),
           @ApiResponse(code = 401, message = "Unauthorized request",        response = Result_Unauthorized.class),
           @ApiResponse(code = 500, message = "Server side Error")
   })
@@ -428,11 +502,11 @@ public class NotificationController extends Controller {
   @Security.Authenticated(Secured_API.class)
   public Result get_unconfirmed_notifications(){
     try{
-      List<Notification> notifications = Notification.find.where().eq("person", SecurityController.getPerson()).eq("confirmation_required", true).eq("confirmed", false).findList();
+      List<Notification> notifications = Notification.find.where().eq("person.id", SecurityController.getPerson().id).eq("notification_importance", Notification_importance.high).eq("confirmed", false).findList();
       if(notifications.isEmpty()) return GlobalResult.result_ok("No new notifications");
 
       for (Notification notification : notifications){
-          NotificationController.send_notification(SecurityController.getPerson(), notification);
+          notification.send(SecurityController.getPerson());
       }
 
       return GlobalResult.result_ok("Notifications were sent again");
@@ -452,7 +526,7 @@ public class NotificationController extends Controller {
       Person person = Person.find.where().eq("mail", mail).findUnique();
       if (person == null) return GlobalResult.notFoundObject("Person not found");
 
-      NotificationController.test_notification(person, help.level, help.importance, help.confirmation_required);
+      NotificationController.test_notification(person, help.level, help.importance, help.type, help.buttons);
       return GlobalResult.result_ok();
 
     }catch (Exception e){
@@ -470,19 +544,39 @@ public class NotificationController extends Controller {
   )
   @ApiResponses(value = {
           @ApiResponse(code = 200, message = "Ok Result",               response = Result_ok.class),
+          @ApiResponse(code = 400, message = "Some Json value Missing", response = Result_JsonValueMissing.class),
+          @ApiResponse(code = 400, message = "Objects not found",       response = Result_NotFound.class),
+          @ApiResponse(code = 400, message = "Something is wrong",      response = Result_BadRequest.class),
           @ApiResponse(code = 401, message = "Unauthorized request",    response = Result_Unauthorized.class),
+          @ApiResponse(code = 403, message = "Need required permission",response = Result_PermissionRequired.class),
           @ApiResponse(code = 500, message = "Server side Error")
   })
   @Security.Authenticated(Secured_API.class)
   public Result notification_confirm(@ApiParam(value = "notification_id String path", required = true) String notification_id){
 
+    final Form<Swagger_Notification_Confirm> form = Form.form(Swagger_Notification_Confirm.class).bindFromRequest();
+    if(form.hasErrors()) {return GlobalResult.formExcepting(form.errorsAsJson());}
+    Swagger_Notification_Confirm help = form.get();
+
     try{
       Notification notification = Notification.find.byId(notification_id);
       if(notification == null) return GlobalResult.notFoundObject("Notification does not exist");
 
-      notification.confirm();
+      if (!notification.confirm_permission()) return GlobalResult.forbidden_Permission();
 
-      return GlobalResult.result_ok();
+      switch (help.action){
+        case "confirm_notification"       : {
+          notification.confirm();
+          return GlobalResult.result_ok("Notification confirmed");
+        }
+        case "accept_project_invitation"  : {
+          return programingPackageController.addParticipantToProject(help.payload, true);
+        }
+        case "reject_project_invitation"  : {
+          return programingPackageController.addParticipantToProject(help.payload, false);
+        }
+        default: return GlobalResult.result_BadRequest("Unknown action");
+      }
 
     }catch (Exception e){
       return Loggy.result_internalServerError(e, request());
